Refactor ClockSeqFunction
