+++ src/main/java/com/github/f4b6a3/uuid/factory/function/impl/DefaultClockSeqFunction.java
@@ -24,9 +24,7 @@
 
 package com.github.f4b6a3.uuid.factory.function.impl;
 
-import static com.github.f4b6a3.uuid.factory.function.ClockSeqFunction.ClockSeqPool.POOL_MIN;
-import static com.github.f4b6a3.uuid.factory.function.ClockSeqFunction.ClockSeqPool.POOL_MAX;
-
+import java.util.SplittableRandom;
 import java.util.concurrent.atomic.AtomicInteger;
 
 import com.github.f4b6a3.uuid.factory.function.ClockSeqFunction;
@@ -35,7 +33,6 @@ import com.github.f4b6a3.uuid.factory.function.ClockSeqFunction;
  * Function that returns a clock sequence.
  * 
  * @see ClockSeqFunction
- * @see ClockSeqFunction.ClockSeqPool
  */
 public final class DefaultClockSeqFunction implements ClockSeqFunction {
 
@@ -71,9 +68,166 @@ public final class DefaultClockSeqFunction implements ClockSeqFunction {
 	 * @return a number
 	 */
 	public int next() {
-		if (this.sequence.incrementAndGet() > POOL_MAX) {
-			this.sequence.set(POOL_MIN);
+		if (this.sequence.incrementAndGet() > ClockSeqPool.POOL_MAX) {
+			this.sequence.set(ClockSeqPool.POOL_MIN);
 		}
 		return this.sequence.updateAndGet(POOL::take);
+	}
+
+	/**
+	 * Nested class that manages a pool of 16384 clock sequence values.
+	 * <p>
+	 * The pool is implemented as an array of 2048 bytes (16384 bits). Each bit of
+	 * the array corresponds to a clock sequence value.
+	 * <p>
+	 * It is used to avoid that two time-based factories use the same clock sequence
+	 * at same time in a class loader.
+	 */
+	static final class ClockSeqPool {
+
+		private final byte[] pool = new byte[2048];
+		private static final int POOL_SIZE = 16384; // 2^14 = 16384
+
+		/**
+		 * The minimum pool size, which is zero.
+		 */
+		public static final int POOL_MIN = 0x00000000;
+
+		/**
+		 * The maximum pool size, which is 16383 (2^14-1).
+		 */
+		public static final int POOL_MAX = 0x00003fff; // 2^14-1 = 16383
+
+		/**
+		 * Take a value from the pool.
+		 * <p>
+		 * If the value to be taken is already in use, it is incremented until a free
+		 * value is found and returned.
+		 * <p>
+		 * In the case that all pool values are in use, the pool is cleared and the last
+		 * incremented value is returned.
+		 * <p>
+		 * It does nothing to negative arguments.
+		 * 
+		 * @param take value to be taken from the pool
+		 * @return the value to be borrowed if not used
+		 */
+		public synchronized int take(final int take) {
+			int value = take;
+			for (int i = 0; i < POOL_SIZE; i++) {
+				if (setBit(value)) {
+					return value;
+				}
+				value = ++value % POOL_SIZE;
+			}
+			clearPool();
+			setBit(value);
+			return value;
+		}
+
+		/**
+		 * Take a random value from the pool.
+		 * 
+		 * @return the random value to be borrowed if not used
+		 */
+		public synchronized int random() {
+			// Choose a random number between 0 and 16383
+			int random = Math.abs(new SplittableRandom().nextInt()) % POOL_SIZE;
+			return this.take(random);
+		}
+
+		/**
+		 * Set a bit from the byte array that represents the pool.
+		 * <p>
+		 * This operation corresponds to setting a value as used.
+		 * <p>
+		 * It returns false if the value is not free.
+		 * <p>
+		 * It does nothing to negative arguments.
+		 * 
+		 * @param value the value to be taken from the pool
+		 * @return true if success
+		 */
+		private synchronized boolean setBit(int value) {
+
+			if (value < 0) {
+				return false;
+			}
+
+			final int byteIndex = value / 8;
+			final int bitIndex = value % 8;
+
+			final int mask = (0x00000001 << bitIndex);
+			final boolean clear = (pool[byteIndex] & mask) == 0;
+
+			if (clear) {
+				pool[byteIndex] = (byte) (pool[byteIndex] | mask);
+				return true;
+			}
+
+			return false;
+		}
+
+		/**
+		 * Check if a value is used out of the pool.
+		 * 
+		 * @param value a value to be checked in the pool
+		 * @return true if the value is used
+		 */
+		public synchronized boolean isUsed(int value) {
+
+			final int byteIndex = value / 8;
+			final int bitIndex = value % 8;
+
+			final int mask = (0x00000001 << bitIndex);
+			boolean clear = (pool[byteIndex] & mask) == 0;
+
+			return !clear;
+		}
+
+		/**
+		 * Check if a value is free in the pool.
+		 * 
+		 * @param value a value to be checked in the pool
+		 * @return true if the value is free
+		 */
+		public synchronized boolean isFree(int value) {
+			return !this.isUsed(value);
+		}
+
+		/**
+		 * Count the used values out of the pool
+		 * 
+		 * @return the count of used values
+		 */
+		public synchronized int countUsed() {
+			int counter = 0;
+			for (int i = 0; i < POOL_SIZE; i++) {
+				if (this.isUsed(i)) {
+					counter++;
+				}
+			}
+			return counter;
+		}
+
+		/**
+		 * Count the free values in the pool.
+		 * 
+		 * @return the count of free values
+		 */
+		public synchronized int countFree() {
+			return POOL_SIZE - this.countUsed();
+		}
+
+		/**
+		 * Clear all bits of the byte array that represents the pool.
+		 * <p>
+		 * This corresponds to marking all pool values as free
+		 */
+		public synchronized void clearPool() {
+			for (int i = 0; i < pool.length; i++) {
+				pool[i] = 0;
+			}
+		}
 	}
 }
