+++ src/test/java/com/github/f4b6a3/uuid/factory/function/impl/DefaultClockSeqFunctionTest.java
@@ -43,7 +43,7 @@ public class DefaultClockSeqFunctionTest {
 		long first = 0;
 		long last = 0;
 
-		int max = ClockSeqFunction.ClockSeqPool.POOL_MAX + 1; // 16,384
+		int max = DefaultClockSeqFunction.ClockSeqPool.POOL_MAX + 1; // 16,384
 		DefaultClockSeqFunction clockSequence = new DefaultClockSeqFunction();
 
 		first = clockSequence.applyAsLong(CURRENT_TIMESTAMP);
@@ -61,7 +61,7 @@ public class DefaultClockSeqFunctionTest {
 		int first = 0;
 		int last = 0;
 
-		int max = ClockSeqFunction.ClockSeqPool.POOL_MAX + 1; // 16,384
+		int max = DefaultClockSeqFunction.ClockSeqPool.POOL_MAX + 1; // 16,384
 		TimeBasedFactory factory = new TimeBasedFactory.Builder().withInstant(Instant.now()).build();
 
 		UUID uuid = factory.create();
@@ -81,7 +81,7 @@ public class DefaultClockSeqFunctionTest {
 	public void testGetClockSequenceUnique1() {
 
 		HashSet<Long> set = new HashSet<>();
-		int max = ClockSeqFunction.ClockSeqPool.POOL_MAX; // 16,383
+		int max = DefaultClockSeqFunction.ClockSeqPool.POOL_MAX; // 16,383
 		DefaultClockSeqFunction clockSequence = new DefaultClockSeqFunction();
 
 		for (int i = 0; i < max; i++) {
@@ -93,7 +93,7 @@ public class DefaultClockSeqFunctionTest {
 	public void testGetClockSequenceUnique2() {
 
 		HashSet<UUID> set = new HashSet<>();
-		int max = ClockSeqFunction.ClockSeqPool.POOL_MAX; // 16,383
+		int max = DefaultClockSeqFunction.ClockSeqPool.POOL_MAX; // 16,383
 		TimeBasedFactory factory = new TimeBasedFactory.Builder().withInstant(Instant.now()).build();
 
 		for (int i = 0; i < max; i++) {
